audit 32-bit and 64-bit system calls
from 'man audit.rules':

When you specify a syscall name, auditctl will look up the name and get
its syscall number. This leads to some problems on bi-arch machines. The
32 and 64 bit syscall numbers sometimes, but not always, line up. So, to
solve this problem, you would generally need to break the rule into 2
with one specifying -F arch=b32 and the other specifying -F arch=b64.
This needs to go in front of the -S option so that auditctl looks at the
right lookup table when returning the number.