audit rules: make ssg benchmark happy
also, from 'man auditctl':

The -w form of writing watches is for backwards
compatibility and the syscall based form is more expressive.
Unlike most syscall auditing rules, watches do not impact
performance based on the number of rules sent to the kernel. The
only valid options when using a watch are the -p and -k. If you
need to anything fancy like audit a specific user accessing a
file, then use the syscall auditing form with the path or dir
fields.